centralize all handlebar helper methods
expand critical request chain extendedInfo so less is computed in its formatter
