+++ lighthouse-core/formatters/partials/critical-request-chains.html
@@ -50,7 +50,7 @@
     <details>
       <summary>View critical network waterfall</summary>
       <div class="initial-nav">Initial navigation</div>
-      {{#createTreeRenderContext this}}
+      {{#createTreeRenderContext this.chains}}
         {{#each this.tree }}
           {{#createContextFor ../tree @key undefined undefined ../startTime ../transferSize }}
             {{> writeNode this }}
